refactor: UITextField 파일 분리
