+++ SOPT_assignment/SOPT_assignment/Extensions/UITextField+.swift
@@ -1,0 +1,16 @@
+//
+//  UITextField+.swift
+//  SOPT_assignment
+//
+//  Created by 이은지 on 4/22/25.
+//
+
+import UIKit.UITextField
+
+extension UITextField {
+    func addLeftPadding() {
+        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: 22, height: self.frame.height))
+        self.leftView = paddingView
+        self.leftViewMode = ViewMode.always
+    }
+}
