+++ SOPT_assignment/SOPT_assignment/LoginViewController.swift
@@ -239,11 +239,3 @@ extension LoginViewController: UITextFieldDelegate {
         textField.layer.borderWidth = 0
     }
 }
-
-extension UITextField {
-    func addLeftPadding() {
-        let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: 22, height: self.frame.height))
-        self.leftView = paddingView
-        self.leftViewMode = ViewMode.always
-    }
-}
